# backend/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TecnoCel Web API",
    "description": "API REST para la tienda en línea de productos tecnológicos TecnoCel Web. Incluye gestión de productos, clientes, ventas, carrito de compras y más.",
    "version": "1.0.0",
    "contact": {
      "name": "Soporte TecnoCel"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Servidor de desarrollo"
    }
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT token. Formato: \"Bearer {token}\""
    },
    "bearerAuthCliente": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT token para clientes. Formato: \"Bearer {token}\""
    }
  },
  "paths": {
...
    "/marcas": {
      "get": {
        "summary": "Obtener todas las marcas",
        "description": "Obtiene todas las marcas activas del sistema",
        "tags": [
          "Marcas"
        ],
        "responses": {
          "200": {
            "description": "Lista de marcas"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      },
      "post": {
        "summary": "Crear marca",
        "description": "Crea una nueva marca en el sistema",
        "tags": [
          "Marcas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "nombre_marca"
                ],
                "properties": {
                  "nombre_marca": {
                    "type": "string"
                  },
                  "descripcion": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Marca creada exitosamente"
          },
          "400": {
            "description": "Datos inválidos"
          },
          "403": {
            "description": "No tiene permisos"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
    "/marcas/{id}": {
      "get": {
        "summary": "Obtener marca por ID",
        "description": "Obtiene una marca específica por su ID",
        "tags": [
          "Marcas"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Marca obtenida"
          },
          "404": {
            "description": "Marca no encontrada"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      },
      "put": {
        "summary": "Actualizar marca",
        "description": "Actualiza una marca existente",
        "tags": [
          "Marcas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre_marca": {
                    "type": "string"
                  },
                  "descripcion": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Marca actualizada exitosamente"
          },
          "400": {
            "description": "Datos inválidos"
          },
          "403": {
            "description": "No tiene permisos"
          },
          "404": {
            "description": "Marca no encontrada"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      },
      "delete": {
        "summary": "Eliminar marca",
        "description": "Elimina (soft delete) una marca del sistema",
        "tags": [
          "Marcas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Marca eliminada exitosamente"
          },
          "403": {
            "description": "No tiene permisos"
          },
          "404": {
            "description": "Marca no encontrada"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
...
    "/ofertas/todas": {
      "get": {
        "summary": "Obtener todas las ofertas (admin)",
        "tags": [
          "Ofertas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de todas las ofertas"
          }
        }
      }
    },
    "/ofertas/producto/{id_producto}": {
      "get": {
        "summary": "Obtener ofertas por producto (admin)",
        "tags": [
          "Ofertas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id_producto",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de ofertas del producto"
          }
        }
      }
    },
    "/ofertas/{id}": {
      "get": {
        "summary": "Obtener oferta por ID (admin)",
        "tags": [
          "Ofertas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalle de oferta"
          }
        }
      },
      "put": {
        "summary": "Actualizar oferta",
        "tags": [
          "Ofertas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Oferta actualizada"
          }
        }
      },
      "delete": {
        "summary": "Eliminar oferta",
        "tags": [
          "Ofertas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Oferta eliminada"
          }
        }
      }
    },
    "/ofertas": {
      "post": {
        "summary": "Crear oferta",
        "tags": [
          "Ofertas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Oferta creada"
          }
        }
      }
    },
    "/ofertas/{id_oferta}/productos": {
      "post": {
        "summary": "Asignar productos a oferta",
        "tags": [
          "Ofertas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id_oferta",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Productos asignados"
          }
        }
      }
    },
    "/ofertas/{id_oferta}/productos/{id_producto}": {
      "delete": {
        "summary": "Quitar producto de oferta",
        "tags": [
          "Ofertas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id_oferta",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "id_producto",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Producto removido"
          }
        }
      }
    },
...
  },
...
}